my-coffeeshop-app/Backend/Twilio/requirements.txt
flask==3.0.0
twilio==9.5.1
python-dotenv==1.0.0